Kuyesera AI Disaster Damage Challenge/requirements.txt
torch==2.4.1
ultralytics==8.3.49
albumentations==1.4.22
pandas==2.2.3
numpy==1.26.4
rasterio==1.4.3
cv2==4.10.0
pillow==11.0.0
ensemble-boxes
shapely==2.0.6
geopandas==1.0.1
pyyaml==6.0.2
tqdm==4.65.2
timm==1.0.12
pyproj==3.6.1
rioxarray==0.15.0
xarray==2024.7.0
boto3==1.36.11
botocore==1.36.3
pystac-client==0.8.3
sat-search==0.3.0
